docs(fe): put interface ProblemDataTop into type.js
