+++ apps/frontend/app/(client)/(main)/contest/[contestId]/@tabs/_components/RenderProblemList.tsx
@@ -1,5 +1,5 @@
 import { DataTable } from '@/app/(client)/(main)/_components/DataTable'
-import type { ProblemDataTop } from '../page'
+import type { ProblemDataTop } from '@/types/type'
 import { RenderProblemColumn } from './RenderProblemColumn'
 
 export function RenderProblemList(
@@ -10,23 +10,23 @@ export function RenderProblemList(
   if (state === 'Upcoming') {
     return (
       <div className="h-[608px] w-[1208px] place-content-center rounded-2xl bg-[#d9d9d940]">
-        <div className="text-center text-xl font-semibold text-[#000000]">
+        <p className="text-center text-xl font-semibold text-[#000000]">
           {`Contest Hasn't Started`}
-        </div>
-        <div className="text-center text-base font-normal text-[#00000080]">
+        </p>
+        <p className="mt-2 text-center text-base font-normal text-[#00000080]">
           The problem list will be released after the start of the contest
-        </div>
+        </p>
       </div>
     )
   } else if (state === 'Ongoing' && !isRegistered) {
     return (
       <div className="h-[608px] w-[1208px] place-content-center rounded-2xl bg-[#d9d9d940]">
-        <div className="text-center text-xl font-semibold text-[#000000]">
+        <p className="text-center text-xl font-semibold text-[#000000]">
           Please Register for The Contest First!
-        </div>
-        <div className="text-center text-base font-normal text-[#00000080]">
+        </p>
+        <p className="mt-2 text-center text-base font-normal text-[#00000080]">
           The problem list only be revealed to contest participants
-        </div>
+        </p>
       </div>
     )
   } else {
@@ -46,7 +46,7 @@ export function RenderProblemList(
             data={transformedProblemData}
             columns={RenderProblemColumn}
             headerStyle={{
-              order: 'text-[#8A8A8A] w-[10px] p-0 md:w-[64px] px-4',
+              order: 'text-[#8A8A8A] w-[10px] p-0 md:w-16 px-4',
               title: 'text-[#8A8A8A] text-left w-[100px] p-0 md:w-[832px] px-6',
               score: 'text-[#8A8A8A] w-4/10 p-0 md:w-[296px] px-4'
             }}
